+++ src/routes/currency/[id]/transactions/+page.svelte
@@ -10,8 +10,8 @@
 {#if currency && transactions && venues}
 	<Page title={currency.name} breadcrumbs={[[`/currency/${currency.id}`, currency.name]]}>
 		{#snippet subtitle()}Transactions{/snippet}
-		<p>These are all {transactions.length} transactions for this currency.</p>
-		<Transactions {transactions} {venues} currencies={[currency]} />
+		<p>These are all {transactions.length} transactions for this currency visible to you.</p>
+		<Transactions {transactions} {venues} currencies={[currency]} testid="currency-transaction" />
 	</Page>
 {:else if currency === null}
 	<h1>Oops.</h1>